Moved from unit-tests.wsjcpp to ctest: test_cycle_dependency
