--- unit-tests.wsjcpp/CMakeLists.user-custom.txt
@@ -1,2 +1,0 @@
-## You can here write some custom includes
-
