--- unit-tests.wsjcpp/build_simple.sh
@@ -1,27 +1,0 @@
-#!/bin/bash
-
-check_ret() {
-    if [ $1 -ne 0 ]; then
-        echo ""
-        echo "!!! FAIL: $2"
-        echo "********************************************************************************"
-        echo ""
-        exit $1
-    else
-        echo ""
-        echo "*** SUCCESS: $2"
-        echo "********************************************************************************"
-        echo ""
-    fi
-}
-
-# if [ ! -d tmp ]; then
-#     mkdir -p tmp
-# fi
-
-cmake -H. -Btmp/release -DCMAKE_BUILD_TYPE=Release
-check_ret $? "configure"
-
-cmake --build tmp/release --config Release
-check_ret $? "make"
-
